+++ app/src/main/java/woowacourse/shopping/presentation/shopping/product/ProductItemListener.kt
@@ -7,7 +7,3 @@ interface ProductItemListener : CartItemListener {
 
     fun loadProducts()
 }
-
-fun interface RecentProductItemListener {
-    fun navigateToDetail(id: Long)
-}
