+++ app/src/main/java/woowacourse/shopping/presentation/shopping/product/RecentProductItemListener.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.presentation.shopping.product
+
+fun interface RecentProductItemListener {
+    fun navigateToDetail(id: Long)
+}
